+++ medil/independence_testing.py
@@ -1,199 +1,13 @@
 """Independence testing on samples of random variables."""
-import numpy as np
-from scipy.spatial.distance import pdist, squareform
+
+from typing import NamedTuple, Optional
+
+from dcor.independence import distance_correlation_t_test
 from multiprocessing import Pool, cpu_count
-
-
-from numpy import linalg as LA
-from scipy.stats import chi2
-
-try:
-    from dcor.independence import distance_correlation_t_test
-    from dcor import pairwise, distance_correlation as dist_corr
-
-    default_measure = "dcor"
-except ImportError:
-    default_measure = "pearson"
-from xicorrelation import xicorr
-
-
-def hypothesis_test(samples, num_resamples, measure=default_measure, alpha=0.05):
-    r"""Performs random permutation tests to estimate independence and
-    returns the estimated Undirected Dependency Graph in the form of an
-    adjacency matrix.
-
-    Parameters
-    ----------
-    samples : 2d numpy array of floats or ints
-              A :math:`M \times N` matrix with :math:`N` samples of
-              :math:`M` random variables.
-
-    num_resamples : int, optional
-                    Number of permutations used to calculate p-values.
-
-    measure : str, optional
-              Measure of association to use as test statistic. Either
-              `pearson` (default if `dcor` package is not installed)
-              for a linear measure or `dcor` (default if installed) for
-              a nonlinear measure.
-
-    alpha : float, optional
-            The threshold on the p-values above which a result is
-            considered statistically significant.
-
-    Returns
-    -------
-    p_values : 2d numpy array of floats
-               A square matrix :math:`P`, where :math:`P_{i,j}` is the
-               probability of obtaining a result at least as extreme
-               as the one given by :math:`N_{i, j}`.
-
-    sample_corr : 2d numpy array of floats
-                  A square matrix :math:`N`, where :math:`N_{i,j}` is
-                  the measured association value (e.g., correlation)
-                  between random variables :math:`R_i` and :math:`R_j`.
-
-    deps : 2d numpy array of bools
-           A square matrix `D`, where :math:`D_{i,j}` is true if and
-           only if the corresponding random variables :math:`R_i` and
-           :math:`R_j` are not estimated to be independent.
-
-    See Also
-    --------
-    pearson_correlation : Used when ``measure == 'pearson'``.
-    distance_correlation : Used when ``measure == 'dcor'``.
-
-    Notes
-    -----
-    The runtime when using ``distance_correlation`` is nearly cut in
-    halve by first generating two permuted samples matrices, ``a`` and
-    ``b`` and then calling ``distance_correlation(a, b)`` and using the
-    upper and lower triangles, as opposed to calling
-    ``distance_correlation(a)`` and ``distance_correlation(b)``
-    separately while using only one triangle from each. A similar trick
-    could be employed in ``pearson_correlation``, but it reduces the
-    runtime by less than 10%. Another (mathematically) similar trick
-    would be to compute the correlation on ``np.vstack((a, b))``, but
-    this also is less than a 10% improvement.
-
-    """
-    if measure == "pearson":
-        compute_corr = pearson_correlation
-    elif measure == "dcor":
-        compute_corr = distance_correlation
-    else:
-        raise ValueError("{} is not a supported measure of association".format(measure))
-
-    sample_corr = compute_corr(samples)
-
-    # initialize aux vars used in loop
-    p_values = np.zeros(sample_corr.shape)
-    num_loops = num_resamples if measure != "dcor" else int(np.ceil(num_resamples / 2))
-    for _ in range(num_loops):
-        perm_corr = compute_corr(samples, perm=True)
-        p_values += np.array(perm_corr >= sample_corr, int)
-
-    p_values += p_values.T
-    p_values /= num_loops if measure != "dcor" else 2 * num_loops
-
-    deps = p_values <= alpha  # reject null hypothesis of independence
-
-    return p_values, sample_corr, deps
-
-
-def distance_correlation(samples, perm=False):
-    r"""Compute distance correlation on samples set (or on permuted samples
-    set, if ``perm`` is ``True``
-
-    Parameters
-    ----------
-    samples : 2d numpy array of floats or ints
-              A :math:`M \times N` matrix with :math:`N` samples of
-              :math:`M` random variables.
-
-    perm : bool, optional
-           Whether distance correlation is computed on permuted or
-           original samples.
-
-    Returns
-    -------
-    2d numpy array of floats
-        A square matrix :math:`C`, where :math:`C_{i,j}` is
-        (if ``perm``, a sample from the null distribution of)
-        the distance correlation between random variables :math:`R_i`
-        and :math:`R_j`.
-
-    """
-    try:
-        pairwise
-    except NameError:
-        raise ImportError(
-            "Install the dcor package to use this feature: `pip install dcor`."
-        )
-    if not perm:
-        with Pool() as pool:
-            corr = pairwise(dist_corr, samples, pool=pool)
-    else:
-        samples_permed = permute_within_rows(samples)
-        samples_permed_2 = permute_within_rows(samples)
-        with Pool() as pool:
-            corr = pairwise(dist_corr, samples_permed, samples_permed_2, pool=pool)
-    return corr
-
-
-def pearson_correlation(samples, perm=False):
-    r"""Computes Pearson product-moment correlation coefficient on samples
-    set (or on permuted samples set, if ``perm`` is ``True``).
-
-    Parameters
-    ----------
-    samples : 2d numpy array of floats or ints
-              A :math:`M \times N` matrix with :math:`N` samples of
-              :math:`M` random variables.
-
-    perm : bool, optional
-           Whether Pearson correlation is computed on permuted or
-           original samples.
-
-    Returns
-    -------
-    2d numpy array of floats
-        A square matrix :math:`C`, where :math:`C_{i,j}` is
-        (if ``perm``, a sample from the null distribution of)
-        the Pearson correlation between random variables :math:`R_i`
-        and :math:`R_j`.
-
-    """
-    if not perm:
-        corr = np.corrcoef(samples)
-    else:
-        samples_permed = permute_within_rows(samples)
-        corr = np.corrcoef(samples, samples_permed)
-    return corr
-
-
-def permute_within_rows(x):
-    """Randomly rearrange values according to column index without
-    changing row index.
-
-    Parameters
-    ----------
-    x : 2d numpy array
-
-    Returns
-    -------
-    2d numpy array
-        Reordered copy of input.
-
-    """
-    # get random new index for col of each element
-    col_idx = np.random.sample(x.shape).argsort(axis=1)
-
-    # keep the row index the same
-    row_idx = np.tile(np.arange(x.shape[0]), (x.shape[1], 1)).T
-
-    # apply the permutaton matrix to permute x
-    return x[row_idx, col_idx]
+import numpy as np
+import numpy.typing as npt
+from scipy.spatial.distance import pdist, squareform
+from scipy.stats import chi2, norm, rankdata
 
 
 def dcov(samples):
@@ -270,3 +84,120 @@ def xicor_test(x_y):
     x, y = x_y
     xi, pvalue = xicorr(x, y)
     return pvalue
+
+
+# The following is a modification of the xicorrelation source code,
+# Copyright 2021 Nikolay Novik (https://github.com/jettify),
+# for compatibility with numpy 2.0+
+class _XiCorr(NamedTuple):
+    xi: float
+    fr: npt.NDArray[np.float64]
+    cu: float
+
+
+class XiCorrResult(NamedTuple):
+    correlation: float
+    pvalue: Optional[float]
+
+
+def _xicorr(x: npt.NDArray[np.float64], y: npt.NDArray[np.float64]) -> _XiCorr:
+    # Ported from original R implementation.
+    # https://github.com/cran/XICOR/blob/master/R/calculateXI.R
+    n = x.size
+    PI = rankdata(x, method="average")
+    fr = rankdata(y, method="average") / n
+    gr = rankdata(-y, method="average") / n
+    cu = np.mean(gr * (1 - gr))
+    A1 = np.abs(np.diff(fr[np.argsort(PI, kind="quicksort")])).sum() / (2 * n)
+    xi = 1.0 - A1 / cu
+    return _XiCorr(xi, fr, cu)
+
+
+def xicorr(x: npt.ArrayLike, y: npt.ArrayLike, ties: bool = True) -> XiCorrResult:
+    """Compute the cross rank increment correlation coefficient xi [1].
+
+    Parameters
+    ----------
+    x, y : array_like
+        Arrays of rankings, of the same shape. If arrays are not 1-D, they
+        will be flattened to 1-D.
+    ties : bool, optional
+        If ties is True, the algorithm assumes that the data has ties and
+        employs the more elaborated theory for calculating s.d. and P-value.
+        Otherwise, it uses the simpler theory. There is no harm in putting
+        ties = True even if there are no ties.
+
+    Returns
+    -------
+    correlation : float
+       The tau statistic.
+    pvalue : float
+       P-values computed by the asymptotic theory.
+
+    See Also
+    --------
+    spearmanr : Calculates a Spearman rank-order correlation coefficient.
+
+
+    Example:
+        >>> from xicorrelation import xicorr
+        >>> x = [1, 2, 3, 4, 5]
+        >>> y = [1, 4, 9, 16, 25]
+        >>> xi, pvalue = xicorr(x, y)
+        >>> print(xi, pvalue)
+
+    References
+    ----------
+    .. [1] Chatterjee, S., "A new coefficient of correlation",
+           https://arxiv.org/abs/1909.10140, 2020.
+
+    Examples
+    --------
+    >>> from scipy import stats
+    >>> x1 = [12, 2, 1, 12, 2]
+    >>> x2 = [1, 4, 7, 1, 0]
+    >>> xi, p_value, _ = xicorr(x1, x2)
+    >>> tau
+    -0.47140452079103173
+    >>> p_value
+    0.2827454599327748
+    """
+    # https://git.io/JSIlN
+    x = np.asarray(x).ravel()
+    y = np.asarray(y).ravel()
+
+    if x.size != y.size:
+        raise ValueError(
+            "All inputs to `xicorr` must be of the same "
+            f"size, found x-size {x.size} and y-size {y.size}"
+        )
+    elif not x.size or not y.size:
+        # Return NaN if arrays are empty
+        return XiCorrResult(np.nan, np.nan)
+
+    r = _xicorr(x, y)
+    xi = r.xi
+    fr = r.fr
+    CU = r.cu
+
+    pvalue: Optional[float] = None
+    # https://git.io/JSIlM
+    n = x.size
+    if not ties:
+        # sd = np.sqrt(2.0 / (5.0 * n))
+        pvalue = 1.0 - norm.cdf(np.sqrt(n) * xi / np.sqrt(2.0 / 5.0))
+    else:
+        qfr = np.sort(fr)
+        ind = np.arange(1, n + 1)
+        ind2 = 2 * n - 2 * ind + 1
+
+        ai = np.mean(ind2 * qfr * qfr) / n
+        ci = np.mean(ind2 * qfr) / n
+        cq = np.cumsum(qfr)
+        m = (cq + (n - ind) * qfr) / n
+        b = np.mean(m**2)
+        v = (ai - 2.0 * b + ci**2) / (CU**2)
+
+        # sd = np.sqrt(v / n)
+        pvalue = 1.0 - norm.cdf(np.sqrt(n) * xi / np.sqrt(v))
+    return XiCorrResult(xi, pvalue)