+++ tests/test_indep_testing.py
@@ -1,5 +1,5 @@
 import numpy as np
-from medil.independence_testing import hypothesis_test
+from medil.independence_testing import estimate_UDG
 
 
 def test_indep_testing():