+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased ##
+- differentiable NCFA
+
+## [1.1.0] - 2024-08-19 ##
+
+### Added ###
+- source code for xicorrelation, updated for NumPy 2.0.0+
+
+### Removed ###
+- `hypothesis_test()` in favor of `estimate_UDG()`
+
 ## [1.0.0] - 2024-05-29 ##
 
 ### Added ###